src/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform
azure-functions
azure-identity>=1.6.0
azure-storage-queue>=12.8.1
pydantic>=1.8.2
pydantic-settings>=2.0.0
azure-monitor-opentelemetry>=1.6.1